+++ app/scripts/dimApp.config.js
@@ -11,13 +11,6 @@
       active: null,
       debug: true
     })
-    .value('dimItemTier', {
-      exotic: 'Exotic',
-      legendary: 'Legendary',
-      rare: 'Rare',
-      uncommon: 'Uncommon',
-      basic: 'Basic'
-    })
     .factory('loadingTracker', ['promiseTracker', function(promiseTracker) {
       return promiseTracker();
     }]);