+++ app/scripts/dimApp.module.js
@@ -4,6 +4,7 @@
   angular.module('dimApp', [
     'ui.router',
     'ngAnimate',
+    'timer',
     'ngAria',
     'ngDialog',
     'ngMessages',