+++ app/scripts/services/dimStoreService.factory.js
@@ -4,28 +4,27 @@
   angular.module('dimApp')
     .factory('dimStoreService', StoreService);
 
-  StoreService.$inject = ['$rootScope', '$q', 'dimBungieService', 'dimPlatformService', 'dimItemTier', 'dimCategory', 'dimItemDefinitions', 'dimBucketService', 'dimStatDefinitions', 'dimObjectiveDefinitions', 'dimTalentDefinitions', 'dimSandboxPerkDefinitions', 'dimYearsDefinitions', 'dimProgressionDefinitions', 'dimRecordsDefinitions', 'dimInfoService', 'SyncService', 'loadingTracker'];
+  StoreService.$inject = ['$rootScope', '$q', 'dimBungieService', 'dimPlatformService', 'dimCategory', 'dimItemDefinitions', 'dimVendorDefinitions', 'dimBucketService', 'dimStatDefinitions', 'dimObjectiveDefinitions', 'dimTalentDefinitions', 'dimSandboxPerkDefinitions', 'dimYearsDefinitions', 'dimProgressionDefinitions', 'dimRecordsDefinitions', 'dimInfoService', 'SyncService', 'loadingTracker'];
 
-  function StoreService($rootScope, $q, dimBungieService, dimPlatformService, dimItemTier, dimCategory, dimItemDefinitions, dimBucketService, dimStatDefinitions, dimObjectiveDefinitions, dimTalentDefinitions, dimSandboxPerkDefinitions, dimYearsDefinitions, dimProgressionDefinitions, dimRecordsDefinitions, dimInfoService, SyncService, loadingTracker) {
+  function StoreService($rootScope, $q, dimBungieService, dimPlatformService, dimCategory, dimItemDefinitions, dimVendorDefinitions, dimBucketService, dimStatDefinitions, dimObjectiveDefinitions, dimTalentDefinitions, dimSandboxPerkDefinitions, dimYearsDefinitions, dimProgressionDefinitions, dimRecordsDefinitions, dimInfoService, SyncService, loadingTracker) {
     var _stores = [];
-    var progressionDefs = {};
-    let recordsDefs = {};
-    var buckets = {};
-    var idTracker = {};
+    var _progressionDefs = {};
+    let _recordsDefs = {};
+    var _buckets = {};
+    var _idTracker = {};
 
-    // A set of items IDs that are new - this is cleared out by the user
-    var _newItems = new Set();
+    var _removedNewItems = new Set();
 
     dimBucketService.then(function(defs) {
-      buckets = defs;
+      _buckets = defs;
     });
     dimProgressionDefinitions.then(function(defs) {
-      progressionDefs = defs;
+      _progressionDefs = defs;
     });
-    dimRecordsDefinitions.then((defs) => { recordsDefs = defs; });
+    dimRecordsDefinitions.then((defs) => { _recordsDefs = defs; });
 
     // A promise used to dedup parallel calls to reloadStores
-    var reloadPromise;
+    var _reloadPromise;
 
     // Cooldowns
     var cooldownsSuperA = ['5:00', '4:46', '4:31', '4:15', '3:58', '3:40'];
@@ -120,6 +119,10 @@
       },
       canBeInLoadout: function() {
         return this.equipment || this.type === 'Material' || this.type === 'Consumable';
+      },
+      // "The Life Exotic" Perk on Exotic Items means you can equip another exotic
+      hasLifeExotic: function() {
+        return this.isExotic && this.talentGrid && (_.find(this.talentGrid.nodes, { name: 'The Life Exotic' }) !== undefined);
       }
     };
 
@@ -136,15 +139,16 @@
       dropNewItem: dropNewItem,
       createItemIndex: createItemIndex,
       processItems: processItems,
-      hasNewItems: Boolean(_newItems.size)
+      hasNewItems: false
     };
 
     $rootScope.$on('dim-active-platform-updated', function() {
       _stores = [];
+      service.hasNewItems = false;
       $rootScope.$broadcast('dim-stores-updated', {
         stores: _stores
       });
-      loadingTracker.addPromise(loadNewItems().then(() => service.reloadStores()));
+      loadingTracker.addPromise(service.reloadStores(true));
     });
 
     return service;
@@ -176,20 +180,41 @@
     // If this is called while a reload is already happening, it'll return the promise
     // for the ongoing reload rather than kicking off a new reload.
     function reloadStores() {
-      if (reloadPromise) {
-        return reloadPromise;
+      const activePlatform = dimPlatformService.getActive();
+      if (_reloadPromise && _reloadPromise.activePlatform === activePlatform) {
+        return _reloadPromise;
       }
 
-      // Save a snapshot of all the items before we update
-      const previousItemsMap = buildItemMap(_stores);
-      const previousItems = new Set(_.keys(previousItemsMap));
+      // Include vendors on the first load and if they're expired
+      const currDate = new Date().toISOString();
+      const includeVendors = !_stores.length || _.any(_stores, (store) => store.minRefreshDate < currDate);
 
-      reloadPromise = dimBungieService.getStores(dimPlatformService.getActive())
-        .then(function(rawStores) {
+      // Save a snapshot of all the items before we update
+      const previousItems = buildItemSet(_stores);
+      const firstLoad = (previousItems.size === 0);
+
+      function fakeItemId(item) {
+        if (activePlatform.fake && item.itemInstanceId !== "0") {
+          item.itemInstanceId = 'fake-' + item.itemInstanceId;
+        }
+      }
+
+      _reloadPromise = $q.all([loadNewItems(activePlatform), dimBungieService.getStores(activePlatform, includeVendors)])
+        .then(function([newItems, rawStores]) {
+          if (activePlatform !== dimPlatformService.getActive()) {
+            throw new Error("Active platform mismatch");
+          }
+
           var glimmer;
           var marks;
+          _removedNewItems.forEach((id) => newItems.delete(id));
+          _removedNewItems.clear();
+          service.hasNewItems = (newItems.size !== 0);
 
-          return $q.all([previousItemsMap, ...rawStores.map(function(raw) {
+          return $q.all([newItems, ...rawStores.map(function(raw) {
+            if (activePlatform !== dimPlatformService.getActive()) {
+              throw new Error("Active platform mismatch");
+            }
             var store;
             var items = [];
             if (!raw) {
@@ -217,7 +242,7 @@
                   if (!sort) {
                     throw new Error("item needs a 'sort' field");
                   }
-                  return buckets[sort].capacity;
+                  return _buckets[sort].capacity;
                 },
                 spaceLeftForItem: function(item) {
                   var sort = item.sort;
@@ -245,6 +270,7 @@
               _.each(raw.data.buckets, function(bucket) {
                 _.each(bucket.items, function(item) {
                   item.bucket = bucket.bucketHash;
+                  fakeItemId(item);
                 });
 
                 items = _.union(items, bucket.items);
@@ -272,18 +298,27 @@
                 percentToNextLevel: raw.character.base.percentToNextLevel / 100.0,
                 progression: raw.character.progression,
                 advisors: raw.character.advisors,
+                vendors: raw.character.vendors,
                 isVault: false
               });
+
+              if (!includeVendors) {
+                var prevStore = _.findWhere(_stores, { id: raw.id });
+                store.vendors = prevStore.vendors;
+                store.minRefreshDate = prevStore.minRefreshDate;
+              }
+
               store.name = store.gender + ' ' + store.race + ' ' + store.class;
 
               store.progression.progressions.forEach(function(prog) {
-                angular.extend(prog, progressionDefs[prog.progressionHash]);
+                angular.extend(prog, _progressionDefs[prog.progressionHash]);
               });
 
               _.each(raw.data.buckets, function(bucket) {
                 _.each(bucket, function(pail) {
                   _.each(pail.items, function(item) {
                     item.bucket = pail.bucketHash;
+                    fakeItemId(item);
                   });
 
                   items = _.union(items, pail.items);
@@ -295,6 +330,7 @@
                   _.each(raw.character.base.inventory.buckets.Invisible, function(pail) {
                     _.each(pail.items, function(item) {
                       item.bucket = pail.bucketHash;
+                      fakeItemId(item);
                     });
 
                     items = _.union(items, pail.items);
@@ -303,7 +339,11 @@
               }
             }
 
-            return processItems(store, items, previousItems).then(function(items) {
+            return processItems(store, items, previousItems, newItems).then(function(items) {
+              if (activePlatform !== dimPlatformService.getActive()) {
+                throw new Error("Active platform mismatch");
+              }
+
               store.items = items;
 
               // by type-bucket
@@ -312,7 +352,7 @@
               });
 
               // Fill in any missing buckets
-              _.values(buckets.byType).forEach(function(bucket) {
+              _.values(_buckets.byType).forEach(function(bucket) {
                 if (!store.buckets[bucket.id]) {
                   store.buckets[bucket.id] = [];
                 }
@@ -322,7 +362,7 @@
                 store.vaultCounts = {};
                 ['Weapons', 'Armor', 'General'].forEach(function(category) {
                   store.vaultCounts[category] = 0;
-                  buckets.byCategory[category].forEach(function(bucket) {
+                  _buckets.byCategory[category].forEach(function(bucket) {
                     if (store.buckets[bucket.id]) {
                       store.vaultCounts[category] += store.buckets[bucket.id].length;
                     }
@@ -330,31 +370,31 @@
                 });
               }
 
-              return store;
+              return processVendors(store.vendors)
+                .then(function(vendors) {
+                  _.each(vendors, function(vendor) {
+                    store.vendors[vendor.vendorHash] = vendor;
+                    if (!store.minRefreshDate || vendor.nextRefreshDate < store.minRefreshDate) {
+                      store.minRefreshDate = vendor.nextRefreshDate;
+                    }
+                  });
+                  return store;
+                });
             });
           })]);
         })
-        .then(function([previousItemsMap, ...stores]) {
+        .then(function([newItems, ...stores]) {
+          if (activePlatform !== dimPlatformService.getActive()) {
+            throw new Error("Active platform mismatch");
+          }
+
           // Save and notify about new items (but only if this wasn't the first load)
-          if (previousItems.size) {
+          if (!firstLoad) {
             // Save the list of new item IDs
-            saveNewItems();
-
-            // Only notify new items from this refresh
-            const notifyNewItems = _.omit(buildItemMap(stores), _.keys(previousItemsMap));
-
-            // Show a toaster about the new items
-            var listStr = '';
-            _.each(notifyNewItems, function(val) {
-              listStr += '<li>[' + val.type + ']' + ' ' + val.name + '</li>';
-            });
-            if (listStr) {
-              dimInfoService.show('newitemsbox', {
-                title: 'New items found',
-                body: '<p>The following items are new:</p><ul>' + listStr + '</ul>',
-                hide: 'Don\'t show me new item notifications'
-              }, 10000);
-            }
+            _removedNewItems.forEach((id) => newItems.delete(id));
+            _removedNewItems.clear();
+            saveNewItems(newItems);
+            service.hasNewItems = (newItems.size !== 0);
           }
 
           _stores = stores;
@@ -365,12 +405,22 @@
 
           return stores;
         })
+        .catch(function(e) {
+          if (e.message === 'Active platform mismatch') {
+            // no problem, just canceling the request
+            return null;
+          }
+          throw e;
+        })
         .finally(function() {
           // Clear the reload promise so this can be called again
-          reloadPromise = null;
+          if (_reloadPromise.activePlatform === activePlatform) {
+            _reloadPromise = null;
+          }
         });
 
-      return reloadPromise;
+      _reloadPromise.activePlatform = activePlatform;
+      return _reloadPromise;
     }
 
     function getStore(id) {
@@ -383,15 +433,15 @@
       var index = item.hash + '-';
       if (item.id === '0') {
         index = index + item.amount;
-        idTracker[index] = (idTracker[index] || 0) + 1;
-        index = index + '-' + idTracker[index];
+        _idTracker[index] = (_idTracker[index] || 0) + 1;
+        index = index + '-' + _idTracker[index];
       } else {
         index = index + item.id;
       }
       return index;
     }
 
-    function processSingleItem(definitions, buckets, statDef, objectiveDef, perkDefs, talentDefs, yearsDefs, progressDefs, previousItems, item, owner) {
+    function processSingleItem(definitions, buckets, statDef, objectiveDef, perkDefs, talentDefs, yearsDefs, progressDefs, previousItems, newItems, item, owner) {
       var itemDef = definitions[item.itemHash];
       // Missing definition?
       if (!itemDef || itemDef.itemName === 'Classified') {
@@ -532,7 +582,7 @@
       // An item is new if it was previously known to be new, or if it's new since the last load (previousItems);
       createdItem.isNew = false;
       try {
-        createdItem.isNew = isItemNew(createdItem.id, previousItems);
+        createdItem.isNew = isItemNew(createdItem.id, previousItems, newItems);
       } catch (e) {
         console.error("Error determining new-ness of " + createdItem.name, item, itemDef, e);
       }
@@ -565,7 +615,7 @@
         try {
           const recordBook = owner.advisors.recordBooks[itemDef.recordBookHash];
 
-          recordBook.records = _.map(_.values(recordBook.records), (record) => _.extend(recordsDefs[record.recordHash], record));
+          recordBook.records = _.map(_.values(recordBook.records), (record) => _.extend(_recordsDefs[record.recordHash], record));
 
           createdItem.objectives = buildRecords(recordBook, objectiveDef);
 
@@ -1053,64 +1103,68 @@
 
     /** New Item Tracking **/
 
-    function buildItemMap(stores) {
-      var itemMap = {};
+    function buildItemSet(stores) {
+      var itemSet = new Set();
       stores.forEach((store) => {
         store.items.forEach((item) => {
-          itemMap[item.id] = { name: item.name, type: item.type };
+          itemSet.add(item.id);
         });
       });
-      return itemMap;
+      return itemSet;
     }
 
     // Should this item display as new? Note the check for previousItems size, so that
     // we don't mark everything as new on the first load.
-    function isItemNew(id, previousItems) {
+    function isItemNew(id, previousItems, newItems) {
       let isNew = false;
-      if (_newItems.has(id)) {
+      if (newItems.has(id)) {
         isNew = true;
+      } else if (_removedNewItems.has(id)) {
+        isNew = false;
       } else if (previousItems.size) {
         // Zero id check is to ignore general items and consumables
         isNew = (id !== '0' && !previousItems.has(id));
         if (isNew) {
-          _newItems.add(id);
+          newItems.add(id);
         }
       }
       return isNew;
     }
 
     function dropNewItem(item) {
-      _newItems.delete(item.id);
-      saveNewItems();
+      _removedNewItems.add(item.id);
       item.isNew = false;
+      loadNewItems(dimPlatformService.getActive()).then((newItems) => {
+        newItems.delete(item.id);
+        service.hasNewItems = (newItems.size !== 0);
+        saveNewItems(newItems);
+      });
     }
 
     function clearNewItems() {
-      _newItems = new Set();
       _stores.forEach((store) => {
         store.items.forEach((item) => {
-          item.isNew = false;
+          if (item.isNew) {
+            _removedNewItems.add(item.id);
+            item.isNew = false;
+          }
         });
       });
       service.hasNewItems = false;
-      saveNewItems();
+      saveNewItems(new Set());
     }
 
-    function loadNewItems() {
-      if (dimPlatformService.getActive()) {
-        _newItems = new Set();
-        service.hasNewItems = false;
+    function loadNewItems(activePlatform) {
+      if (activePlatform) {
         return SyncService.get().then(function processCachedNewItems(data) {
-          _newItems = new Set(data[newItemsKey()]);
-          service.hasNewItems = Boolean(_newItems.size);
+          return new Set(data[newItemsKey()]);
         });
       }
-      return $q.when();
+      return $q.resolve(new Set());
     }
 
-    function saveNewItems() {
-      service.hasNewItems = Boolean(_newItems.size);
-      SyncService.set({ [newItemsKey()]: [..._newItems] });
+    function saveNewItems(newItems) {
+      SyncService.set({ [newItemsKey()]: [...newItems] });
     }
 
     function newItemsKey() {
@@ -1118,8 +1172,8 @@
       return 'newItems-' + (platform ? platform.type : '');
     }
 
-    function processItems(owner, items, previousItems = new Set()) {
-      idTracker = {};
+    function processItems(owner, items, previousItems = new Set(), newItems = new Set()) {
+      _idTracker = {};
       return $q.all([
         dimItemDefinitions,
         dimBucketService,
@@ -1129,7 +1183,8 @@
         dimTalentDefinitions,
         dimYearsDefinitions,
         dimProgressionDefinitions,
-        previousItems])
+        previousItems,
+        newItems])
         .then(function(args) {
           var result = [];
           _.each(items, function(item) {
@@ -1207,6 +1262,49 @@
       } else {
         return '-:--';
       }
+    }
+
+    function processVendors(vendors) {
+      return $q.all([dimVendorDefinitions, dimItemDefinitions])
+        .then(function([vendorDefs, itemDefs]) {
+          return $q.all(_.map(vendors, function(vendor, vendorHash) {
+            var def = vendorDefs[vendorHash];
+            vendor.vendorName = def.vendorName;
+            vendor.vendorIcon = def.factionIcon || def.vendorIcon;
+            vendor.items = [];
+            vendor.costs = [];
+            if (vendor.enabled) {
+              var items = [];
+              _.each(vendor.saleItemCategories, function(categoryData) {
+                var filteredSaleItems = _.filter(categoryData.saleItems, function(saleItem) {
+                  return saleItem.item.isEquipment && saleItem.costs.length;
+                });
+                items.push(...filteredSaleItems);
+              });
+              vendor.items = _.pluck(items, 'item');
+
+              var costs = _.reduce(items, function(o, saleItem) {
+                o[saleItem.item.itemHash] = {
+                  cost: saleItem.costs[0].value,
+                  currency: _.pick(itemDefs[saleItem.costs[0].itemHash], 'itemName', 'icon', 'itemHash')
+                };
+                return o;
+              }, {});
+              vendor.costs = costs;
+            }
+            return processItems({ id: null }, vendor.items)
+              .then(function(items) {
+                vendor.items = {};
+                vendor.items.armor = _.filter(items, function(item) {
+                  return item.primStat && item.primStat.statHash === 3897883278 && item.primStat.value >= 280;
+                });
+                vendor.items.weapons = _.filter(items, function(item) {
+                  return item.primStat && item.primStat.statHash === 368428387 && item.primStat.value >= 280;
+                });
+                return vendor;
+              });
+          }));
+        });
     }
 
     function getStatsData(data) {